% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-wordcloud.R
\name{custom_wordcloud}
\alias{custom_wordcloud}
\title{Custom Word Cloud with Color Palette}
\usage{
custom_wordcloud(words, freq, palette = "Set3")
}
\arguments{
\item{words}{Vector of words}

\item{freq}{Vector of frequencies}

\item{palette}{RColorBrewer palette (default is "Set3")}
}
\value{
A word cloud plot
}
\description{
This function generates a word cloud with a selected color palette.
}
